+++ grunt/options/copy.js
@@ -6,9 +6,17 @@
 var config = require('../config');
 
 module.exports = {
-  build: {
+  js: {
     files: [{
       src: config.js.files,
+      dest: config.destDir
+    }]
+  },
+  root: {
+    files: [{
+      expand: true,
+      cwd: config.srcDir,
+      src: '*.{png,xml,ico,svg,txt}',
       dest: config.destDir
     }]
   }