+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ### HEAD
+* Copy files that are needed in dest via grunt copy
 * Use grunt-sass with libsass for speed
 * Upgrade `console.log` / `console.warn` (#112)
 * Remove timestamp logging when running Karma